tests/wpt/tests/fetch/metadata/generated/element-input-image.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-input-image.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTML "input" element with type="button"</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, test) {
    const input = document.createElement('input');
    input.setAttribute('type', 'image');

    document.body.appendChild(input);
    test.add_cleanup(() => input.remove());

    return new Promise((resolve) => {
        input.onload = input.onerror = resolve;
        input.setAttribute('src', url);
      });
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - Not sent to non-trustworthy same-origin destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - Not sent to non-trustworthy same-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - Not sent to non-trustworthy cross-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-mode');
        });
  }, 'sec-fetch-mode - Not sent to non-trustworthy same-origin destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-mode');
        });
  }, 'sec-fetch-mode - Not sent to non-trustworthy same-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-mode');
        });
  }, 'sec-fetch-mode - Not sent to non-trustworthy cross-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-dest');
        });
  }, 'sec-fetch-dest - Not sent to non-trustworthy same-origin destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-dest');
        });
  }, 'sec-fetch-dest - Not sent to non-trustworthy same-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-dest');
        });
  }, 'sec-fetch-dest - Not sent to non-trustworthy cross-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - Not sent to non-trustworthy same-origin destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - Not sent to non-trustworthy same-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - Not sent to non-trustworthy cross-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
        });
  }, 'sec-fetch-storage-access - Not sent to non-trustworthy same-origin destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
        });
  }, 'sec-fetch-storage-access - Not sent to non-trustworthy same-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
        });
  }, 'sec-fetch-storage-access - Not sent to non-trustworthy cross-site destination, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin', 'httpOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - HTTPS downgrade (header not sent), no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpOrigin', 'httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - HTTPS upgrade, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin', 'httpOrigin', 'httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade, no attributes');
  </script>
  </body>
</html>
